// src/redux/adminLicenses/action-types.ts
export enum ActionTypes {
    ADD_ANSWER = "ADD_ANSWER",
    ADD_FIELDS = "ADD_FIELDS",
    ADD_IMAGE = "ADD_IMAGE",
    ADD_QUESTION = "ADD_QUESTION",
    ADD_SECTION = "ADD_SECTION",
    DELETE_ANSWER = "DELETE_ANSWER",
    DELETE_IMAGE = "DELETE_IMAGE",
    DELETE_QUESTION = "DELETE_QUESTION",
    DELETE_SECTION = "DELETE_SECTION",
    DELETE_SLIDE_PROP = "DELETE_SLIDE_PROP",
    EDIT_ANSWER = "EDIT_ANSWER",
    EDIT_QUESTION = "EDIT_QUESTION",
    EDIT_SECTION = "EDIT_SECTION",
    GET_SEARCH_LICENSE = "GET_SEARCH_LICENSE",
    RESET_ADMIN_LICENSE_STATE = "RESET_ADMIN_LICENSE_STATE",
    SET_GET_LICENSE_ERROR = "SET_GET_LICENSE_ERROR"
}